+++ scripts/build-app.sh
@@ -111,9 +111,9 @@ cat > "$APP/Contents/Info.plist" <<'PLIST'
   <key>CFBundlePackageType</key>
   <string>APPL</string>
   <key>CFBundleShortVersionString</key>
-  <string>1.0.5</string>
+  <string>1.0.6</string>
   <key>CFBundleVersion</key>
-  <string>15</string>
+  <string>16</string>
   <key>LSMinimumSystemVersion</key>
   <string>12.0</string>
   <key>NSHighResolutionCapable</key>